+++ client/src/components/ContentControl/DataBlockEditForm/DataBlockEditForm.js
@@ -1,216 +1,29 @@
-import React, { useState, createContext } from 'react'
+import _ from 'lodash'
+import React, { useState } from 'react'
 import update from 'immutability-helper'
-import { DragDropContext, Draggable, Droppable } from 'react-beautiful-dnd'
-import _, { create } from 'lodash'
-import { EditorState, convertToRaw, ContentState } from 'draft-js'
-import htmlToDraft from 'html-to-draftjs'
-import draftToHtml from 'draftjs-to-html'
+import { DragDropContext, Droppable } from 'react-beautiful-dnd'
 
 import { FaPlus } from 'react-icons/fa'
-
 import Accordion from './Accordion/Accordion'
 import TextField from './TextField/TextField'
 import RichTextField from './RichTextField/RichTextField'
 import Label from './Label/Label'
 
+import {
+    createNewDataObject,
+    editorStateToHtml,
+    generateFormData,
+    generateFormObject,
+    generateUpdateFromPath,
+    htmlToEditorState,
+    reorder
+} from './helpers'
+
 import "./PageEditForm.css"
 
-function htmlToEditorState(html) {
-    const { contentBlocks, entityMap} = htmlToDraft(html)
-    return EditorState.createWithContent(
-        ContentState.createFromBlockArray(
-            contentBlocks, entityMap
-        )
-    )
-}
 
-function editorStateToHtml(editorState) {
-    return draftToHtml(
-        convertToRaw(
-            editorState.getCurrentContent()
-        )
-    )
-}
-
-function generateFormObject(formData, formStructure) {
-    const form = formData
-    const newForm = {}
-    let meta = formStructure._meta
-    if (!meta) {
-        meta = {
-            quantity: 1
-        }
-    }
-
-    newForm._accordionTitle = meta.title ? formData[meta.title] : "DataBlock"
-
-    Object.keys(formData).forEach((key) => {
-        if (key === '_meta' || key == '_accordionTitle') {
-            return
-        }
-        
-        // const contentState = ContentState.createFromBlockArray(contentBlocks, entityMap);
-
-        let val = formData[key]
-        
-        if (typeof val === 'string') {
-            let valueType
-            if (typeof formStructure[key] === 'string') {
-                valueType = formStructure[key]
-            }
-            else {
-                valueType = formStructure[key].contentType
-            }
-
-            let value = val
-            if (valueType === 'richtext') {
-                value = htmlToEditorState(value)
-            }
-
-            const newVal = {
-                _value: value,
-                _type: valueType
-            }
-
-            newForm[key] = newVal
-        }
-        else if (typeof val === 'object') {
-            if (!Array.isArray(val)) {
-                val = [val]
-            }
-
-            const newVal = []
-            val.forEach((formVal) => (
-                newVal.push(generateFormObject(formVal, formStructure[key]))
-            ))
-            newForm[key] = newVal
-        }
-    })
-
-    return newForm
-}
-
-function generateFormData(formObject, formStructure) {
-    const formData = {}
-    if (!formStructure._meta) {
-        formStructure._meta = {
-            quantity: 1
-        }
-    }
-    
-    Object.entries(formObject).forEach(([key, val]) => {
-        if (key === '_accordionTitle') {
-            return
-        }
-
-        // Checking like this because we wanna allow empty strings
-        if (val._value !== undefined && val._value !== null) {
-            let value = val._value
-            if (val._type === 'richtext') {
-                value = editorStateToHtml(value)
-                console.log('value is', value)
-            }
-            formData[key] = value
-        }
-        else {
-            if (formStructure[key]._meta.quantity === 1
-                || formStructure[key]._meta.quantity === undefined) {
-                    formData[key] = generateFormData(val[0], formStructure[key])
-            }
-            else {
-                const vals = []
-                val.forEach((v) => vals.push(generateFormData(v, formStructure[key])))
-                formData[key] = vals
-            }
-        }
-    })
-
-    return formData
-}
-
-function generateUpdateFromPath(path, updateVal) {
-    const keys = path.split('.')
-    const updateObj = {}
-    let curr = updateObj
-
-    keys.forEach((key) => {
-        curr[key] = {}
-        curr = curr[key]
-    })
-    
-    curr["$set"] = updateVal
-    return updateObj
-}
-
-function reorder(list, startIndex, endIndex) {
-    const result = Array.from(list)
-    const [removed] = result.splice(startIndex, 1)
-    result.splice(endIndex, 0, removed)
-
-    return result
-}
-
-function createNewDataObject(template) {
-    const result = {}
-    if (!template._meta) {
-        template._meta = {
-            quantity: 1
-        }
-    }
-
-    Object.entries(template).forEach(([key, val]) => {
-        if (key === '_meta') {
-            return
-        }
-
-        if (typeof val === "string") {
-            let value = "bruh"
-            if (val === 'richtext') {
-                value = htmlToEditorState(value)
-            }
-            result[key] = {
-                _type: val,
-                _value: value
-            }
-            return
-        }
-        else if (typeof val === 'object' && val.contentType) {
-            let value = 'bruh'
-            if (val.contentType === 'richtext') {
-                value = htmlToEditorState(value)
-            }
-            result[key] = {
-                _type: val.contentType,
-                _value: value
-            }
-        }
-        else {
-            const quantity = val._meta.quantity
-            let iterations = 1
-            if (typeof quantity === 'number' && quantity !== -1) {
-                iterations = quantity
-            }
-            else {
-                if (quantity.min) {
-                    iterations = quantity.min
-                }
-            }
-            const newSubData = []
-            for (let i = 0; i < iterations; i++) {
-                newSubData.push(createNewDataObject(val))
-            }
-            result[key] = newSubData
-        }
-    })
-
-    result._accordionTitle = template._meta.title ? result[template._meta.title]._value : "DataBlock"
-
-    return result
-}
-
-
-const PageEditForm = (props) => {
-    const { data, template } = props
+const DataBlockEditForm = (props) => {
+    const { data, template, loading } = props
     const formData = generateFormObject(data, template)
 
     const [formState, setFormState] = useState(formData)
@@ -232,6 +45,7 @@ const PageEditForm = (props) => {
                 index={props.index}
                 level={0}
                 root
+                loading={loading}
             />
     )
 }
@@ -418,6 +232,24 @@ const ContentBlockForm = (props) => {
         }
     })
 
+    let saveButton = (
+        <div
+            className="FormSaveButton"
+            onClick={props.onSave}
+        >
+            Save
+        </div>
+    )
+    if (props.loading) {
+        saveButton = (
+            <div
+                className='FormSaveButton'
+            >
+                Loading...
+            </div>
+        )
+    }
+
     const content = (
         <div style={{backgroundColor: "#fff"}}>
             <div
@@ -430,13 +262,8 @@ const ContentBlockForm = (props) => {
             </div>
             {
                 props.root
-                ?   <div
-                        className="FormSaveButton"
-                        onClick={props.onSave}
-                    >
-                        Save
-                    </div>
-                : null
+                ?   saveButton
+                :   null
             }
             {
                 deleteable
@@ -467,51 +294,4 @@ const ContentBlockForm = (props) => {
     )
 }
 
-export default PageEditForm
-
-
-// const template = {
-//     _meta: {
-//         quantity: -1,
-//         title: "button_title"
-//     },
-//     button_title: "text",
-//     content_title: "text",
-//     content: "richtext",
-//     additional: {
-//         _meta: {
-//             quantity: 2,
-//             title: "interest"
-//         },
-//         interest: "text",
-//         papers: "text"
-//     }
-// }
-
-// const data = {
-//     button_title: "Academics",
-//     content_title: "Academics Content",
-//     content: "These are my academic interests - I have no academic interests",
-//     additional: [
-//         {
-//             interest: "SDN",
-//             papers: "5"
-//         },
-//         {
-//             interest: "AI/ML",
-//             papers: "0"
-//         }
-//     ]
-// }
-
-// const testData = generateFormObject(data, template)
-// console.log(testData)
-// const original = generateFormData(testData, template)
-// console.log(original)
-// console.log(JSON.stringify(testData))
-
-
-// const path = 'info.name.first'
-// const val = 'dushyant'
-// const obj = generateUpdateFromPath(path, val)
-// console.log(JSON.stringify(obj))
+export default DataBlockEditForm
