+++ client/src/components/ContentControl/PageGroup/Controls/SharedDataControl/SharedDataControl.js
@@ -1,27 +1,50 @@
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { connect } from 'react-redux'
 
 import api from '../../../../../axios'
+import * as actions from '../../../../../store/actions/index'
+
 import { MdLibraryBooks } from 'react-icons/md'
 import Control from '../../../../Control/Control'
-import PageEditForm from '../../../PageEditForm/PageEditForm'
+import DataBlockEditForm from '../../../DataBlockEditForm/DataBlockEditForm'
+
+import { generateFormData, generateFormObject } from '../../../DataBlockEditForm/helpers'
+
+function getFormObject(dataBlock) {
+    if (!dataBlock || !dataBlock.data) {
+        return null
+    }
+    if (!dataBlock.template || !dataBlock.template.structure) {
+        return null
+    }
+
+    return generateFormObject(dataBlock.data, dataBlock.template.structure)
+}
 
 const SharedDataControl = (props) => {
     const [state, setState] = useState({
-        loading: false
+        loading: false,
+        formState: getFormObject(props.pageGroup.dataBlock)
     })
-    
+
+    console.log('STATE IS', state)
+
     let comp = (
         <div>
             This PageGroup has no Shared Data
         </div>
     )
 
+    useEffect(() => {
+        console.log('in useEffect')
+    }, [props.pageGroup])
+
     function handleSave(data) {
         setState({...state, loading: true})
         api.post(`/pageGroup/updateData/${props.pageGroup._id}`, {data})
             .then((response) => {
                 console.log('OK SAVED')
+                props.setContent(data)
                 setState({...state, loading: false})
             })
             .catch((e) => {
@@ -29,13 +52,20 @@ const SharedDataControl = (props) => {
                 setState({...state, loading: false})
             })
     }
+
+    function handleUpdate(formState) {
+        return {...state, formState: formState}
+    }
     
     if (props.pageGroup.dataBlock && props.pageGroup.dataBlock.template) {
         comp = (
-            <PageEditForm
+            <DataBlockEditForm
+                formState={state.formState}
+                onUpdate={(formState) => handleUpdate(formState)}
                 data={props.pageGroup.dataBlock.data}
                 template={props.pageGroup.dataBlock.template.structure}
                 onSave={(data) => handleSave(data)}
+                loading={state.loading}
             />
         )
     }
@@ -65,4 +95,10 @@ const mapStateToProps = (state) => {
     }
 }
 
-export default connect(mapStateToProps)(SharedDataControl)
+const mapDispatchToProps = (dispatch) => {
+    return {
+        setContent: (data) => dispatch(actions.setContent({ data, type: 'PAGE_GROUP' }))
+    }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(SharedDataControl)
