+++ client/src/components/ControlBar/ControlBar.css
@@ -1,7 +1,8 @@
 .ControlBar {
     background-color: var(--entity-color);
     box-shadow: var(--light-shadow);
-    width: 80px;
+    min-width: 80px;
+    /* float: right; */
     /* width: 100vw; */
     /* height: 100vh; */
 }
